test: cover the notification card animation styles in their own file
Name the file after what it covers, next to `animation.test.js`, which covers
the animation lifecycle of the mixin with an animation of its own.

Shorten the animation in the tests that only assert that it is applied, and wait
for events instead of a timeout that has to outlast the default duration. One
test asserts the default duration and delay.
